initial materialize design

+++ embercli/app/routes/posts/show.js
@@ -3,8 +3,11 @@ import Ember from 'ember';
 export default Ember.Route.extend({
   actions: {
     deleteComment: function(comment) {
-      comment.deleteRecord();
-      comment.save();
+      if(confirm('Are you sure?')) {
+      	//comment.destroyRecord();
+      	comment.deleteRecord();
+        comment.save();
+      }
     }
   }
 });
